refactor(components): refactor component exports
closes #73
