+++ src/components/atoms/Icon/index.ts
@@ -1,0 +1,2 @@
+export * from './Icon';
+export { default } from './Icon';
