+++ src/components/organisms/SelectWallet/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SelectWallet';
