+++ src/components/templates/Layout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Layout';
